Update active-directory-cloud-provisioning-sso.md
Modified Step 2.2. It previously said to go to Program Files but that only works if AD Connect is installed. The purpose of Step 1 is only to extract the msi not install AD Connect, so that folder does not exist. Instead I changed 2.2 to direct the user to the extract folder from Step 1